forgot to fix protocols to conform to new led_functions

+++ device.h
@@ -40,7 +40,7 @@ struct corsair_device_info {
 };
 
 extern struct corsair_device_info corsairlink_devices[];
-extern size_t corsairlink_device_count;
+extern uint8_t corsairlink_device_count;
 
 #endif
